# Optimizing month_to_season function