The curent version of month_to_season in [climatology.py](https://github.com/NCAR/geocat-comp/blob/main/src/geocat/comp/climatology.py) calculates a seasonal mean by calculating the mean for all four seasons versus just the user-specified season of interest. This results in unnecessary computations.

In an effort to improve the efficiency of this function, I have implemented a version that filters the data before computation.

These are the results of the comparison test when ran on a Casper computation node with 1 CPU with 10 GB of memory.
```
---Test 6 million data points---
Old Method
0.11941519472748041 seconds
New Method
0.05952004995197058 seconds

Change:  50.15705489757617 %


---Test 60 million data points---
Old Method
0.7728856382891536 seconds
New Method
0.6625110507011414 seconds

Change:  14.280843390017639 %


---Test 600 million data points---
Old Method
10.698792199604213 seconds
New Method
8.26200267020613 seconds

Change:  22.776304875687085 %
```
The average performance improvement for 600 million data points averaged over 30 trials with randomized data was 25.56%.