@hCraker 2021-05-27
Performance update:
```
---Test 6 million data points---
Old Method
0.5723645305261016 seconds
New Method
0.033852167427539825 seconds

Change:  94.08555813260759 %


---Test 60 million data points---
Old Method
0.5565323093906045 seconds
New Method
0.2899643750861287 seconds

Change:  47.89801594742345 %


---Test 600 million data points---
Old Method
6.964924789965153 seconds
New Method
3.12193295173347 seconds
```
Average over 30 trials for 600 million data points: 57.21% faster